requirements.txt
crewai>=0.11.0
langchain-google-genai>=0.0.5
python-dotenv==1.0.0
langchain>=0.1.0
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
google-generativeai==0.3.2
typing==3.7.4.3 